Took non-test functions out of testing script
